feat(database): update BaseUtility and enhance user relationships
Update the version of the `BaseUtility` package to `7.1.0` in both
`MyDiet.Auth.Infrastructure.csproj` and `MyDiet.Core.Infrastructure.csproj`.
Add a `UserId` property to the `CalendarMeal`, `Meal`, `MealFood`,
`MealSwap`, and `Plan` classes to establish relationships with the
`CoreUser` class. Update the `OnModelCreating` method in
`MyDietCoreDbContext` to include foreign key relationships for the
new `UserId` property. Modify SQL schema files to include the
`UserId` column and foreign key constraints. Change the DSP in
`MyDiet.Shared.Sql.sqlproj` to support Azure SQL Database.
Add a new snapshot file for the project.